smart casting in codegen from int -> double, double -> int

+++ data/vectors/sum_vector.R
@@ -1,4 +1,4 @@
-add_vec <- function(): vector<int> {
+add_vec <- function(): vector<double> {
     x <- c(1, 2, 3)
     y <- c(4, 5, 6)
